+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added permissions for automation account to handle PSPs and Silences
+- Added Silences Automation account
 
 ## [0.28.1] - 2022-05-18